backend/pyproject.toml
[project]
name = "bebrivus-backend"
version = "0.1.0"
description = "beBrivus - AI-powered opportunity platform backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0.0",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-cors-headers>=4.3.0",
    "django-extensions>=3.2.0",
    "python-decouple>=3.8",
    "pillow>=10.0.0",
    "django-filter>=23.3",
    "django-taggit>=4.0.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dateutil>=2.8.0",
    "django-cleanup>=8.0.0",
    "django-storages>=1.14.0",
    "psycopg2-binary>=2.9.0",
    "google-generativeai>=0.8.5",
    "drf-nested-routers>=0.95.0",
    "channels>=4.0.0",
    "channels-redis>=4.1.0",
]
